Add update user path.

+++ server/config.go
@@ -12,10 +12,10 @@ import (
 )
 
 const (
-	defaultCSRFMaxAge       int64  = 60 * 60 * 24    // 1 day in seconds
-	defaultSessionMaxAge    int64  = 60 * 60 * 24    // 1 day in seconds
-	defaultReadTimeout      int64  = 5               // In seconds
-	defaultWriteTimeout     int64  = 60              // In seconds
+	defaultCSRFMaxAge       uint32 = 60 * 60 * 24    // 1 day in seconds
+	defaultSessionMaxAge    uint32 = 60 * 60 * 24    // 1 day in seconds
+	defaultReadTimeout      uint32 = 5               // In seconds
+	defaultWriteTimeout     uint32 = 60              // In seconds
 	defaultReqBodySizeLimit int64  = 3 * 1024 * 1024 // 3 MiB
 	defaultPluginBatchLimit uint32 = 20
 	defaultListen                  = "4443"
@@ -26,11 +26,11 @@ type Config struct {
 	HTTPSCert        string // File path
 	HTTPSKey         string // File path
 	CSRFKey          string // File path
-	CSRFMaxAge       int64
+	CSRFMaxAge       uint32
 	SessionKey       string // File path
-	SessionMaxAge    int64
-	ReadTimeout      int64
-	WriteTimeout     int64
+	SessionMaxAge    uint32
+	ReadTimeout      uint32
+	WriteTimeout     uint32
 	ReqBodySizeLimit int64
 	PluginBatchLimit uint32
 	Listen           string
